contrib/omnihub/requirements.txt
keras-applications
huggingface-hub
tensorflow-hub
onnx
requests
pytest
torch-model-archiver
cchardet
tensorflow
transformers
scikit-learn